+++ blueprints/incredible-squaring-eigenlayer/aggregator/src/lib.rs
@@ -1,0 +1,12 @@
+use alloy_sol_types::sol;
+use serde::{Deserialize, Serialize};
+
+pub mod aggregator;
+
+sol!(
+    #[allow(missing_docs)]
+    #[sol(rpc)]
+    #[derive(Debug, Serialize, Deserialize)]
+    IncredibleSquaringTaskManager,
+    "../contracts/out/IncredibleSquaringTaskManager.sol/IncredibleSquaringTaskManager.json"
+);