+++ cli/src/deploy.rs
@@ -2,9 +2,7 @@ use alloy_provider::network::TransactionBuilder;
 use alloy_provider::Provider;
 pub use alloy_signer_local::PrivateKeySigner;
 use color_eyre::eyre::{self, Context, ContextCompat, OptionExt, Result};
-use gadget_blueprint_proc_macro_core::{
-    JobResultVerifier, ServiceBlueprint, ServiceRegistrationHook, ServiceRequestHook,
-};
+use gadget_blueprint_proc_macro_core::{BlueprintManager, ServiceBlueprint};
 use gadget_sdk::clients::tangle::runtime::TangleConfig;
 pub use k256;
 use std::fmt::Debug;
@@ -54,10 +52,12 @@ pub async fn generate_service_blueprint<P: Into<PathBuf>, T: AsRef<str>>(
         .exec()
         .context("Getting Metadata about the workspace")?;
 
-    let package = find_package(&metadata, pkg_name)?;
-    let mut blueprint = load_blueprint_metadata(package)?;
-    build_contracts_if_needed(package, &blueprint).context("Building contracts")?;
-    deploy_contracts_to_tangle(rpc_url.as_ref(), package, &mut blueprint, signer_evm).await?;
+    let package = find_package(&metadata, pkg_name)?.clone();
+    let package_clone = &package.clone();
+    let mut blueprint =
+        tokio::task::spawn_blocking(move || load_blueprint_metadata(&package)).await??;
+    build_contracts_if_needed(package_clone, &blueprint).context("Building contracts")?;
+    deploy_contracts_to_tangle(rpc_url.as_ref(), package_clone, &mut blueprint, signer_evm).await?;
 
     bake_blueprint(blueprint)
 }
@@ -112,7 +112,9 @@ pub async fn deploy_to_tangle(
     Ok(event.blueprint_id)
 }
 
-pub fn load_blueprint_metadata(package: &cargo_metadata::Package) -> Result<ServiceBlueprint> {
+pub fn load_blueprint_metadata(
+    package: &cargo_metadata::Package,
+) -> Result<ServiceBlueprint<'static>> {
     let blueprint_json_path = package
         .manifest_path
         .parent()
@@ -142,32 +144,16 @@ async fn deploy_contracts_to_tangle(
     signer_evm: Option<PrivateKeySigner>,
 ) -> Result<()> {
     enum ContractKind {
-        RegistrationHook,
-        RequestHook,
-        JobVerifier(usize),
+        Manager,
     }
     let rpc_url = rpc_url.replace("ws", "http").replace("wss", "https");
-    let mut contract_paths = Vec::new();
-    match blueprint.registration_hook {
-        ServiceRegistrationHook::None => {}
-        ServiceRegistrationHook::Evm(ref path) => {
-            contract_paths.push((ContractKind::RegistrationHook, path))
+    let contract_paths = match blueprint.manager {
+        BlueprintManager::Evm(ref path) => vec![(ContractKind::Manager, path)],
+        _ => {
+            eprintln!("Unsupported blueprint manager kind");
+            vec![]
         }
     };
-
-    match blueprint.request_hook {
-        ServiceRequestHook::None => {}
-        ServiceRequestHook::Evm(ref path) => contract_paths.push((ContractKind::RequestHook, path)),
-    };
-
-    for (id, job) in blueprint.jobs.iter().enumerate() {
-        match job.verifier {
-            JobResultVerifier::None => {}
-            JobResultVerifier::Evm(ref path) => {
-                contract_paths.push((ContractKind::JobVerifier(id), path))
-            }
-        };
-    }
 
     let abs_contract_paths: Vec<_> = contract_paths
         .into_iter()
@@ -225,16 +211,8 @@ async fn deploy_contracts_to_tangle(
                 alloy_network::ReceiptResponse::contract_address(&receipt).unwrap();
             eprintln!("Contract {name} deployed at: {contract_address}");
             match kind {
-                ContractKind::RegistrationHook => {
-                    blueprint.registration_hook =
-                        ServiceRegistrationHook::Evm(contract_address.to_string());
-                }
-                ContractKind::RequestHook => {
-                    blueprint.request_hook = ServiceRequestHook::Evm(contract_address.to_string());
-                }
-                ContractKind::JobVerifier(id) => {
-                    blueprint.jobs[*id].verifier =
-                        JobResultVerifier::Evm(contract_address.to_string());
+                ContractKind::Manager => {
+                    blueprint.manager = BlueprintManager::Evm(contract_address.to_string());
                 }
             }
         } else {
@@ -250,24 +228,13 @@ fn build_contracts_if_needed(
     package: &cargo_metadata::Package,
     blueprint: &ServiceBlueprint,
 ) -> Result<()> {
-    let mut pathes_to_check = Vec::new();
-
-    match blueprint.registration_hook {
-        ServiceRegistrationHook::None => {}
-        ServiceRegistrationHook::Evm(ref path) => pathes_to_check.push(path),
+    let pathes_to_check = match blueprint.manager {
+        BlueprintManager::Evm(ref path) => vec![path],
+        _ => {
+            eprintln!("Unsupported blueprint manager kind");
+            vec![]
+        }
     };
-
-    match blueprint.request_hook {
-        ServiceRequestHook::None => {}
-        ServiceRequestHook::Evm(ref path) => pathes_to_check.push(path),
-    };
-
-    for job in blueprint.jobs.iter() {
-        match job.verifier {
-            JobResultVerifier::None => {}
-            JobResultVerifier::Evm(ref path) => pathes_to_check.push(path),
-        };
-    }
 
     let abs_pathes_to_check: Vec<_> = pathes_to_check
         .into_iter()