+++ sdk/src/keystore/backend/mem.rs
@@ -4,8 +4,11 @@ use alloc::collections::BTreeMap;
 use alloc::string::ToString;
 use alloc::sync::Arc;
 use alloc::vec::Vec;
+use core::str::FromStr;
 use w3f_bls::SerializableToBytes;
 
+use crate::keystore::bn254::Public;
+use crate::keystore::ecdsa::Secret;
 use crate::keystore::{bls381, bn254, ecdsa, ed25519, sr25519, Backend, Error};
 
 /// The type alias for the In Memory `KeyMap`.
@@ -169,6 +172,14 @@ impl<RwLock: lock_api::RawRwLock> Backend for InMemoryKeystore<RwLock> {
         Ok(public)
     }
 
+    fn ecdsa_generate_from_string(&self, string: &str) -> Result<ecdsa::Public, Error> {
+        let secret = Secret::from_slice(hex::decode(string).unwrap().as_slice()).unwrap();
+        let public = secret.public_key();
+        let old = self.ecdsa.write().insert(public, secret);
+        assert!(old.is_none(), "generated key already exists");
+        Ok(public)
+    }
+
     fn ecdsa_sign(
         &self,
         public: &ecdsa::Public,
@@ -209,8 +220,24 @@ impl<RwLock: lock_api::RawRwLock> Backend for InMemoryKeystore<RwLock> {
     }
 
     fn bls_bn254_generate_new(&self, seed: Option<&[u8]>) -> Result<bn254::Public, Error> {
-        let secret = bn254::generate_with_optional_seed(seed);
-        let public = bn254::to_public(&secret);
+        let secret = bn254::generate_with_optional_seed(seed)?;
+        let pair = eigensdk::crypto_bls::BlsKeyPair::new(secret.to_string())
+            .map_err(|e| Error::BlsBn254(e.to_string()))?;
+        let public = pair.public_key();
+        let old = self
+            .bn254
+            .write()
+            .insert(BlsBn254PublicWrapper(public.clone()), secret);
+        assert!(old.is_none(), "generated key already exists");
+        Ok(public)
+    }
+
+    fn bls_bn254_generate_from_string(&self, secret: String) -> Result<Public, Error> {
+        let pair = eigensdk::crypto_bls::BlsKeyPair::new(secret.clone())
+            .map_err(|e| Error::BlsBn254(e.to_string()))?;
+        let public = pair.public_key();
+        let secret = bn254::Secret::from_str(&secret)
+            .map_err(|_| Error::BlsBn254("Invalid BLS BN254 secret".to_string()))?;
         let old = self
             .bn254
             .write()
@@ -244,6 +271,14 @@ impl<RwLock: lock_api::RawRwLock> Backend for InMemoryKeystore<RwLock> {
     fn expose_ecdsa_secret(&self, public: &ecdsa::Public) -> Result<Option<ecdsa::Secret>, Error> {
         let lock = self.ecdsa.read();
         Ok(lock.get(public).cloned())
+    }
+
+    fn get_ecdsa_signer_string(&self, public: &ecdsa::Public) -> Result<String, Error> {
+        let read_secret = self
+            .expose_ecdsa_secret(public)?
+            .ok_or(Error::Ecdsa("Failed to expose secret".to_string()))?;
+        let hex_secret = hex::encode(read_secret.to_bytes().as_slice());
+        Ok(hex_secret)
     }
 
     fn expose_ed25519_secret(