+++ blueprint-metadata/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.3](https://github.com/tangle-network/gadget/compare/blueprint-metadata-v0.1.2...blueprint-metadata-v0.1.3) - 2024-10-23
+
+### Fixed
+
+- *(sdk)* updated keystore support and fixes ([#368](https://github.com/tangle-network/gadget/pull/368))
+
 ## [0.1.2](https://github.com/tangle-network/gadget/compare/blueprint-metadata-v0.1.1...blueprint-metadata-v0.1.2) - 2024-09-24
 
 ### Other