+++ sdk/CHANGELOG.md
@@ -7,6 +7,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0](https://github.com/tangle-network/gadget/compare/gadget-sdk-v0.1.1...gadget-sdk-v0.2.0) - 2024-10-23
+
+### Added
+
+- expose bootnodes in GadgetConfiguration ([#366](https://github.com/tangle-network/gadget/pull/366))
+- *(sdk)* re-export `libp2p`
+- add optional data dir to blueprint manager ([#342](https://github.com/tangle-network/gadget/pull/342))
+- eigenlayer incredible squaring blueprint and test ([#312](https://github.com/tangle-network/gadget/pull/312))
+- allow env vars for `ContextConfig` args ([#339](https://github.com/tangle-network/gadget/pull/339))
+
+### Fixed
+
+- *(sdk)* updated keystore support and fixes ([#368](https://github.com/tangle-network/gadget/pull/368))
+- *(sdk)* [**breaking**] downgrade substrate dependencies for now
+- add `data_dir` back to `GadgetConfiguration` ([#350](https://github.com/tangle-network/gadget/pull/350))
+
+### Other
+
+- release ([#378](https://github.com/tangle-network/gadget/pull/378))
+- release ([#362](https://github.com/tangle-network/gadget/pull/362))
+- Fix eigenlayer example ([#375](https://github.com/tangle-network/gadget/pull/375))
+- update to latest changes in tangle ([#367](https://github.com/tangle-network/gadget/pull/367))
+- Event Workflows (phase 1: Custom listeners) ([#359](https://github.com/tangle-network/gadget/pull/359))
+- Multi job runner + SDK main macro ([#346](https://github.com/tangle-network/gadget/pull/346))
+- Event Listener Upgrade + Wrapper Types + sdk::main macro ([#333](https://github.com/tangle-network/gadget/pull/333))
+- update naming ([#343](https://github.com/tangle-network/gadget/pull/343))
+- Fix typos ([#329](https://github.com/tangle-network/gadget/pull/329))
+- Event listener ([#317](https://github.com/tangle-network/gadget/pull/317))
+- Return output from cmd exec ([#328](https://github.com/tangle-network/gadget/pull/328))
+- release ([#314](https://github.com/tangle-network/gadget/pull/314))
+
 ## [0.1.2](https://github.com/tangle-network/gadget/compare/gadget-sdk-v0.1.1...gadget-sdk-v0.1.2) - 2024-09-30
 
 ### Other